// src/main/resources/angular/src/app/api/model/vehicle.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VehicleRegistrationState } from './vehicleRegistrationState';
import { VehicleType } from './vehicleType';
import { VehicleDocumentation } from './vehicleDocumentation';
import { AvailablePermission } from './availablePermission';


export interface Vehicle { 
    vehicleId?: number;
    /**
     * You need to specify or there will be specified guildMemberId or userId.
     */
    guildMemberId?: number;
    guildId?: number;
    /**
     * You need to specify or there will be specified guildMemberId or userId.
     */
    userId?: number;
    plateNumber: string;
    alias?: string;
    imageData?: string;
    international: boolean;
    /**
     * timestamp of vehicle creation.
     */
    registrationDate?: Date;
    /**
     * timestamp of vehicle deletion.
     */
    closingDate?: Date;
    ecologicalTag?: string;
    vehicleType?: VehicleType;
    vehicleRegistrationState?: VehicleRegistrationState;
    /**
     * Specified by bo user when registration state is changed.
     */
    descriptionState?: string;
    model?: string;
    marca?: string;
    guildName?: string;
    guildMemberName?: string;
    permisAreaVerda?: boolean;
    permisAreaDUM?: boolean;
    vehicleDocumentation?: Array<VehicleDocumentation>;
    availablePermission?:Array<AvailablePermission>;
    numUsers?:number;

}