// src/main/resources/angular/src/app/api/model/vehicleRegistrationState.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: rbarbado@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { I18n } from './i18n';


export interface VehicleRegistrationState { 
    vehicleRegistrationStateId: number;
    name?: string;
    i18n?: I18n;
}